+++ src/jobs/update-attribute/resync-attribute-floor-sell.ts
@@ -1,0 +1,102 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
+
+import _ from "lodash";
+import { Job, Queue, QueueScheduler, Worker } from "bullmq";
+import { randomUUID } from "crypto";
+
+import { logger } from "@/common/logger";
+import { redis, redlock } from "@/common/redis";
+import { config } from "@/config/index";
+
+import { idb } from "@/common/db";
+import * as resyncAttributeCache from "@/jobs/update-attribute/resync-attribute-cache";
+import { fromBuffer } from "@/common/utils";
+
+const QUEUE_NAME = "resync-attribute-floor-value-queue";
+
+export const queue = new Queue(QUEUE_NAME, {
+  connection: redis.duplicate(),
+  defaultJobOptions: {
+    attempts: 10,
+    removeOnComplete: 10000,
+    removeOnFail: 10000,
+  },
+});
+new QueueScheduler(QUEUE_NAME, { connection: redis.duplicate() });
+
+// BACKGROUND WORKER ONLY
+if (config.doBackgroundWork) {
+  const worker = new Worker(
+    QUEUE_NAME,
+    async (job: Job) => {
+      const { continuation } = job.data;
+      const limit = 500;
+      let continuationFilter = "";
+
+      if (continuation != "") {
+        continuationFilter = `WHERE id > '${continuation}'`;
+      }
+
+      const query = `SELECT id
+                     FROM collections
+                     ${continuationFilter}
+                     ORDER BY id ASC
+                     LIMIT ${limit}`;
+
+      const collections = await idb.manyOrNone(query);
+
+      if (collections) {
+        const collectionsIds = _.join(
+          _.map(collections, (collection) => collection.id),
+          "','"
+        );
+
+        const tokensQuery = `
+            SELECT DISTINCT ON (key, value) key, value, tokens.contract, tokens.token_id
+            FROM collections
+            JOIN tokens ON collections.contract = tokens.contract
+            JOIN token_attributes ON tokens.contract = token_attributes.contract AND token_attributes.token_id = tokens.token_id
+            WHERE collections.id IN ('$/collectionsIds:raw/')
+            AND tokens.floor_sell_value IS NOT NULL
+        `;
+
+        const tokens = await idb.manyOrNone(tokensQuery, { collectionsIds });
+
+        _.forEach(tokens, (token) => {
+          resyncAttributeCache.addToQueue(fromBuffer(token.contract), token.token_id, 0);
+        });
+
+        job.data.cursor = null;
+        if (_.size(collections) == limit) {
+          const lastCollection = _.last(collections);
+          job.data.cursor = lastCollection.id;
+        }
+      }
+    },
+    { connection: redis.duplicate(), concurrency: 3 }
+  );
+
+  worker.on("completed", async (job) => {
+    if (job.data.cursor) {
+      logger.info(QUEUE_NAME, `Updated up to lastCollection=${job.data.cursor}`);
+      await addToQueue(job.data.cursor);
+    }
+  });
+
+  worker.on("error", (error) => {
+    logger.error(QUEUE_NAME, `Worker errored: ${error}`);
+  });
+
+  redlock
+    .acquire(["sell-value-resync1"], 60 * 60 * 24 * 30 * 1000)
+    .then(async () => {
+      await addToQueue();
+    })
+    .catch(() => {
+      // Skip on any errors
+    });
+}
+
+export const addToQueue = async (continuation = "") => {
+  await queue.add(randomUUID(), { continuation });
+};