+++ src/migrations/1646736672602_attributes.sql
@@ -49,6 +49,9 @@ CREATE INDEX "attributes_collection_id_floor_sell_value_index"
 CREATE INDEX "attributes_collection_id_top_buy_value_index"
   ON "attributes" ("collection_id" ASC NULLS LAST, "top_buy_value" DESC NULLS LAST);
 
+  CREATE INDEX attributes_key_collection_id_floor_sell_value_index
+    ON attributes (key ASC NULLS LAST, collection_id ASC NULLS LAST, floor_sell_value DESC NULLS LAST);
+
 ALTER TABLE "attributes"
   ADD CONSTRAINT "attributes_pk"
   PRIMARY KEY ("id");